Feat: 1.Hanging done. 2.Fixed the repo.
