+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/SensorGoBildaPinpointExample.java
@@ -22,6 +22,7 @@
 
 package org.firstinspires.ftc.teamcode.test;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
@@ -58,6 +59,7 @@ For support, contact tech@gobilda.com
 -Ethan Doak
  */
 
+@Disabled
 @TeleOp(name="goBILDA® PinPoint Odometry Example", group="Linear OpMode")
 //@Disabled
